Add logger interceptor. Build order methods. Prepare DTOs

+++ src/products/products.controller.ts
@@ -1,6 +1,8 @@
-import { Controller, Delete, Get, Post, Param, Body } from '@nestjs/common';
+import { Controller, Delete, Get, Post, Put, Param, Body, NotFoundException } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { ParseUUIDPipe } from '@nestjs/common';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -12,12 +14,17 @@ export class ProductsController {
     }
 
     @Get('/:id')
-    getById(@Param('id') id: string) {
-        return this.productsService.getById(id);
+    getById(@Param('id', new ParseUUIDPipe()) id: string) {
+        const product = this.productsService.getById(id);
+        if(!product) throw new NotFoundException('Product not found');
+        return product;
     }
 
     @Delete('/:id')
-    deleteById(@Param('id') id: string) {
+    deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+        if(!this.productsService.getById(id)) {
+            throw new NotFoundException('Product not found');
+        }
         this.productsService.deleteById(id);
         return { success: true };
     }
@@ -25,6 +32,15 @@ export class ProductsController {
     @Post('/')
     create(@Body() productData: CreateProductDTO) {
         return this.productsService.create(productData);
+    }
+
+    @Put('/:id')
+    updateById(@Param('id', new ParseUUIDPipe()) id: string, @Body() productData: UpdateProductDTO) {
+        if(!this.productsService.getById(id)) {
+            throw new NotFoundException('Product not found');
+        }
+        this.productsService.updateById(id, productData);
+        return { success: true}
     }
 
 }
